requirement.txt

discord.py==2.6.4
dotenv==0.9.9
PyNaCl==1.6.0
python-dotenv==1.1.1
yt-dlp==2025.10.22
